+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * 3.11.0
     * Add ability to retrieve foldered metadata (Reports, Dashboards, Email templates, and Documents) thanks @kenhuman!
         * You can currently only save reports and dashboards. Some email template metadata can be saved, but the templates themselves can't be saved.
+    * Added a query editor to replace the simple SOQL/TOQL query menu options
 * 3.10.3
     * Update Slack invite link
 * 3.10.2